Spltting up the repo
- All the vsix related files will be moved to a separate repo: azPipeline-KubernetesUI-devopsExtension
https://github.com/Microsoft/azPipeline-KubernetesUI-devopsExtension
