+++ LICENSE.txt
@@ -1,6 +1,8 @@
-The MIT License (MIT)
+azpipelines-kubernetesUI
 
 Copyright (c) Microsoft Corporation. All rights reserved.
+
+The MIT License (MIT)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
