--- azure-devops-extension.json
@@ -1,30 +1,0 @@
-{
-    "manifestVersion": 1,
-    "id": "azdevops-kube-summary",
-    "publisher": "",
-    "version": "1.0.17",
-    "name": "Kubernetes Summary",
-    "description": "kubernetes summary",
-    "public": false,
-    "categories": [
-        "Azure Pipelines"
-    ],
-    "targets": [
-        {
-            "id": "Microsoft.VisualStudio.Services"
-        }
-    ],
-    "icons": {
-        "default": "logo.png"
-    },
-    "files": [
-        {
-            "path": "static",
-            "addressable": true
-        },
-        {
-            "path": "dist",
-            "addressable": true
-        }
-    ]
-}
