--- src/Common.scss
@@ -1,5 +1,0 @@
-#root {
-    height: 100%;
-    width: 100%;
-    display: flex;
-}
