--- src/WebUI/KubeSummary.html
@@ -1,11 +1,0 @@
-<!--
-    Copyright (c) Microsoft Corporation. All rights reserved.
-    Licensed under the MIT license.
--->
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-    <body>
-        <div id="root"></div>
-        <script type="text/javascript" src="WebUI/WebUI.js" charset="utf-8"></script>
-    </body>
-</html>
