--- src/WebUI/KubeSummary.json
@@ -1,21 +1,0 @@
-{
-    "contributions": [
-        {
-            "id": "kube-summary-hub",
-            "type": "ms.vss-web.hub",
-            "targets": [
-                "ms.vss-build-web.build-release-hub-group"
-            ],
-            "includes": [
-            ],
-            "properties": {
-                "name": "Kubernetes Summary",
-                "uri": "dist/KubeSummary.html",
-                "icon": "asset://static/sample-icon.png"
-            }
-        }
-    ],
-    "scopes": [
-        "vso.serviceendpoint_manage"
-    ]
-}
