--- webpack.config.js
@@ -1,56 +1,0 @@
-const path = require("path");
-const fs = require("fs");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-
-// Webpack entry points. Mapping from resulting bundle name to the source file entry.
-const entries = {};
-
-// Loop through subfolders in the "src" folder and add an entry for each one
-const srcDir = path.join(__dirname, "src");
-fs.readdirSync(srcDir).filter(dir => {
-    if (fs.statSync(path.join(srcDir, dir)).isDirectory()) {
-        entries[dir] = "./" + path.relative(process.cwd(), path.join(srcDir, dir, dir));
-    }
-});
-
-module.exports = {
-    entry: entries,
-    output: {
-        filename: "[name]/[name].js"
-    },
-    resolve: {
-        extensions: [".ts", ".tsx", ".js"],
-    },
-    stats: {
-        warnings: false
-    },
-    module: {
-        rules: [
-            {
-                test: /\.tsx?$/,
-                loader: "ts-loader"
-            },
-            {
-                test: /\.scss$/,
-                use: ["style-loader", "css-loader", "azure-devops-ui/buildScripts/css-variables-loader", "sass-loader"]
-            },
-            {
-                test: /\.css$/,
-                use: ["style-loader", "css-loader"],
-            },
-            {
-                test: /\.woff$/,
-                use: [{
-                    loader: 'base64-inline-loader'
-                }]
-            },
-            {
-                test: /\.html$/,
-                loader: "file-loader"
-            }
-        ]
-    },
-    plugins: [
-        new CopyWebpackPlugin([ { from: "**/*.html", context: "src/WebUI" }])
-    ]
-};
